+++ .eslintrc
@@ -18,7 +18,6 @@
     "no-bitwise": 2,
     "no-caller": 2,
     "no-catch-shadow": 2,
-    "no-comma-dangle": 0,
     "no-cond-assign": [2, "except-parens"],
     "no-constant-condition": 2,
     "no-continue": 0,
@@ -31,7 +30,6 @@
     "no-duplicate-case": 2,
     "no-else-return": 0,
     "no-empty": 2,
-    "no-empty-class": 0,
     "no-empty-character-class": 2,
     "no-empty-label": 2,
     "no-eq-null": 0,
@@ -40,9 +38,8 @@
     "no-extend-native": 2,
     "no-extra-bind": 2,
     "no-extra-boolean-cast": 2,
-    "no-extra-parens": 0,
+    "no-extra-parens": 2,
     "no-extra-semi": 2,
-    "no-extra-strict": 2,
     "no-fallthrough": 2,
     "no-floating-decimal": 0,
     "no-func-assign": 2,
@@ -82,9 +79,8 @@
     "no-process-env": 0,
     "no-process-exit": 0,
     "no-proto": 2,
-    "no-redeclare": 0,
+    "no-redeclare": 2,
     "no-regex-spaces": 2,
-    "no-reserved-keys": 0,
     "no-restricted-modules": 0,
     "no-return-assign": 0,
     "no-script-url": 2,
@@ -92,7 +88,6 @@
     "no-sequences": 2,
     "no-shadow": 0,
     "no-shadow-restricted-names": 2,
-    "no-space-before-semi": 0,
     "no-spaced-func": 2,
     "no-sparse-arrays": 2,
     "no-sync": 0,
@@ -111,14 +106,14 @@
     "no-unused-vars": 2,
     "no-use-before-define": [2, "nofunc"],
     "no-void": 0,
-    "no-var": 0,
-    "prefer-const": 0,
+    "no-var": 2,
+    "no-const-assign": 2,
+    "prefer-const": 2,
     "no-warning-comments": [0, {
       "terms": ["todo", "fixme", "xxx"],
       "location": "start"
     }],
     "no-with": 2,
-    "no-wrap-func": 2,
     "array-bracket-spacing": [0, "never"],
     "accessor-pairs": 0,
     "block-scoped-var": 0,
@@ -132,7 +127,7 @@
     "consistent-return": 0,
     "consistent-this": [0, "that"],
     "constructor-super": 0,
-    "curly": [2, "all"],
+    "curly": [2, "multi-line"],
     "default-case": 0,
     "dot-location": 0,
     "dot-notation": 0,
@@ -140,13 +135,11 @@
     "eqeqeq": 2,
     "func-names": 0,
     "func-style": [0, "declaration"],
-    "generator-star": 0,
     "generator-star-spacing": 0,
-    "global-strict": [0, "never"],
     "guard-for-in": 0,
     "handle-callback-err": 0,
-    "indent": 0,
-    "key-spacing": [0, {
+    "indent": [2, 2],
+    "key-spacing": [2, {
       "beforeColon": false,
       "afterColon": true
     }],
@@ -174,12 +167,9 @@
       "after": true
     }],
     "sort-vars": 0,
-    "space-after-function-name": [0, "never"],
     "space-after-keywords": [0, "always"],
     "space-before-blocks": [0, "always"],
     "space-before-function-paren": [0, "always"],
-    "space-before-function-parentheses": [0, "always"],
-    "space-in-brackets": [0, "never"],
     "space-in-parens": [0, "never"],
     "space-infix-ops": 2,
     "space-return-throw-case": 2,
@@ -188,8 +178,7 @@
       "nonwords": false
     }],
     "spaced-comment": 0,
-    "spaced-line-comment": [0, "always"],
-    "strict": 0,
+    "strict": [2, "global"],
     "use-isnan": 2,
     "valid-jsdoc": 0,
     "valid-typeof": 0,
