+++ .travis.yml
@@ -3,13 +3,12 @@ os:
 language: node_js
 
 node_js:
-  - "0.12"
-  - "4"
-  - "iojs-1"
   - "iojs-2"
+  - "iojs-3"
+  - "4"
 
 install:
-  - npm install spm coveralls mocha
+  - npm install spm coveralls
 
 before_script:
   - node_modules/spm/bin/spm-install
@@ -17,8 +16,7 @@ before_script:
 
 script:
   - node_modules/spm/bin/spm-test
-  - node_modules/.bin/mocha -R spec tests/detector-spec.js
-  - node_modules/.bin/mocha -R spec tests/morerule-test.js
+  - node_modules/.bin/mocha -R spec tests/*.js
 
 after_success:
-  - node_modules/spm/bin/spm-test --coveralls | node_modules/.bin/coveralls
+  - cat coverage/lcov.info | node_modules/.bin/coveralls
