+++ lib/web-rules.js
@@ -1,33 +1,17 @@
+"use strict";
 
-var detector = {};
-var NA_VERSION = "-1";
-var win = this;
-var external;
-var re_msie = /\b(?:msie |ie |trident\/[0-9].*rv[ :])([0-9.]+)/;
-var re_blackberry_10 = /\bbb10\b.+?\bversion\/([\d.]+)/;
-var re_blackberry_6_7 = /\bblackberry\b.+\bversion\/([\d.]+)/;
-var re_blackberry_4_5 = /\bblackberry\d+\/([\d.]+)/;
+const win = typeof window === "undefined" ? global : window;
+const external = win.external;
+const re_msie = /\b(?:msie |ie |trident\/[0-9].*rv[ :])([0-9.]+)/;
+const re_blackberry_10 = /\bbb10\b.+?\bversion\/([\d.]+)/;
+const re_blackberry_6_7 = /\bblackberry\b.+\bversion\/([\d.]+)/;
+const re_blackberry_4_5 = /\bblackberry\d+\/([\d.]+)/;
 
-function toString(object){
-  return Object.prototype.toString.call(object);
-}
-function isObject(object){
-  return toString(object) === "[object Object]";
-}
-function isFunction(object){
-  return toString(object) === "[object Function]";
-}
-function each(object, factory){
-  for(var i = 0, l = object.length; i < l; i++){
-    if(factory.call(object, object[i], i) === false){
-      break;
-    }
-  }
-}
+const NA_VERSION = "-1";
 
 // 硬件设备信息识别表达式。
 // 使用数组可以按优先级排序。
-var DEVICES = [
+const DEVICES = [
   ["nokia", function(ua){
     // 不能将两个表达式合并，因为可能出现 "nokia; nokia 960"
     // 这种情况下会优先识别出 nokia/-1
@@ -69,8 +53,8 @@ var DEVICES = [
       /\bm([0-9cx]{1,4})\b/;
   }],
   ["nexus", /\bnexus ([0-9s.]+)/],
-  ["huawei", function(ua){
-    var re_mediapad = /\bmediapad (.+?)(?= build\/huaweimediapad\b)/;
+  ["huawei", function(ua) {
+    const re_mediapad = /\bmediapad (.+?)(?= build\/huaweimediapad\b)/;
     if(ua.indexOf("huawei-huawei") !== -1){
       return /\bhuawei\-huawei\-([a-z0-9\-]+)/;
     }else if(re_mediapad.test(ua)){
@@ -118,7 +102,7 @@ var DEVICES = [
 ];
 
 // 操作系统信息识别表达式
-var OS = [
+const OS = [
   ["wp", function(ua){
     if(ua.indexOf("windows phone ") !== -1){
       return /\bwindows phone (?:os )?([0-9.]+)/;
@@ -159,56 +143,12 @@ var OS = [
   ["windowsce", /\bwindows ce(?: ([0-9.]+))?/],
   ["symbian", /\bsymbian(?:os)?\/([0-9.]+)/],
   ["blackberry", function(ua){
-    var m = ua.match(re_blackberry_10) ||
+    const m = ua.match(re_blackberry_10) ||
       ua.match(re_blackberry_6_7) ||
       ua.match(re_blackberry_4_5);
     return m ? {version: m[1]} : "blackberry";
   }],
 ];
-
-// 解析使用 Trident 内核的浏览器的 `浏览器模式` 和 `文档模式` 信息。
-// @param {String} ua, userAgent string.
-// @return {Object}
-function IEMode(ua){
-  if(!re_msie.test(ua)){ return null; }
-
-  var m,
-      engineMode, engineVersion,
-      browserMode, browserVersion;
-
-  // IE8 及其以上提供有 Trident 信息，
-  // 默认的兼容模式，UA 中 Trident 版本不发生变化。
-  if(ua.indexOf("trident/") !== -1){
-    m = /\btrident\/([0-9.]+)/.exec(ua);
-    if (m && m.length >= 2) {
-      // 真实引擎版本。
-      engineVersion = m[1];
-      var v_version = m[1].split(".");
-      v_version[0] = parseInt(v_version[0], 10) + 4;
-      browserVersion = v_version.join(".");
-    }
-  }
-
-  m = re_msie.exec(ua);
-  browserMode = m[1];
-  var v_mode = m[1].split(".");
-  if (typeof browserVersion === "undefined") {
-    browserVersion = browserMode;
-  }
-  v_mode[0] = parseInt(v_mode[0], 10) - 4;
-  engineMode = v_mode.join(".");
-  if (typeof engineVersion === "undefined") {
-    engineVersion = engineMode;
-  }
-
-  return {
-    browserVersion: browserVersion,
-    browserMode: browserMode,
-    engineVersion: engineVersion,
-    engineMode: engineMode,
-    compatible: engineVersion !== engineMode,
-  };
-}
 
 // 针对同源的 TheWorld 和 360 的 external 对象进行检测。
 // @param {String} key, 关键字，用于检测浏览器的安装路径中出现的关键字。
@@ -218,19 +158,19 @@ function checkTW360External(key){
   try{
     //        360安装路径：
     //        C:%5CPROGRA~1%5C360%5C360se3%5C360SE.exe
-    var runpath = external.twGetRunPath.toLowerCase();
+    const runpath = external.twGetRunPath.toLowerCase();
     // 360SE 3.x ~ 5.x support.
     // 暴露的 external.twGetVersion 和 external.twGetSecurityID 均为 undefined。
     // 因此只能用 try/catch 而无法使用特性判断。
-    var security = external.twGetSecurityID(win);
-    var version = external.twGetVersion(security);
+    const security = external.twGetSecurityID(win);
+    const version = external.twGetVersion(security);
 
     if (runpath && runpath.indexOf(key) === -1) { return false; }
     if (version){return {version: version}; }
   }catch(ex){ /* */ }
 }
 
-var ENGINE = [
+const ENGINE = [
   ["edgehtml", /edge\/([0-9.]+)/],
   ["trident", re_msie],
   ["blink", function(){
@@ -238,8 +178,8 @@ var ENGINE = [
   }],
   ["webkit", /\bapplewebkit[\/]?([0-9.+]+)/],
   ["gecko", function(ua){
-    var match;
-    if ((match = ua.match(/\brv:([\d\w.]+).*\bgecko\/(\d+)/))) {
+    const match = ua.match(/\brv:([\d\w.]+).*\bgecko\/(\d+)/);
+    if (match) {
       return {
         version: match[1] + "." + match[2],
       };
@@ -251,7 +191,7 @@ var ENGINE = [
   ["u2", /\bu2\/([0-9.]+)/],
   ["u3", /\bu3\/([0-9.]+)/],
 ];
-var BROWSER = [
+const BROWSER = [
   // Microsoft Edge Browser, Default browser in Windows 10.
   ["edge", /edge\/([0-9.]+)/],
   // Sogou.
@@ -268,13 +208,13 @@ var BROWSER = [
   // 只能通过 UA 和程序安装路径中的应用程序名来区分。
   // TheWorld 的 UA 比 360 更靠谱，所有将 TheWorld 的规则放置到 360 之前。
   ["theworld", function(){
-    var x = checkTW360External("theworld");
+    const x = checkTW360External("theworld");
     if(typeof x !== "undefined"){ return x; }
     return "theworld";
   }],
   // 360SE, 360EE.
   ["360", function(ua) {
-    var x = checkTW360External("360se");
+    const x = checkTW360External("360se");
     if(typeof x !== "undefined"){ return x; }
     if(ua.indexOf("360 aphone browser") !== -1){
       return /\b360 aphone browser \(([^\)]+)\)/;
@@ -301,7 +241,7 @@ var BROWSER = [
       return /\bliebaofast\/([0-9.]+)/;
     }
     if(ua.indexOf("lbbrowser") === -1){ return false; }
-    var version;
+    let version;
     try{
       if(external && external.LiebaoGetVersion){
         version = external.LiebaoGetVersion();
@@ -321,8 +261,8 @@ var BROWSER = [
   ["mi", /\bmiuibrowser\/([0-9.]+)/],
   // Opera 15 之后开始使用 Chromniun 内核，需要放在 Chrome 的规则之前。
   ["opera", function(ua){
-    var re_opera_old = /\bopera.+version\/([0-9.ab]+)/;
-    var re_opera_new = /\bopr\/([0-9.]+)/;
+    const re_opera_old = /\bopera.+version\/([0-9.ab]+)/;
+    const re_opera_new = /\bopr\/([0-9.]+)/;
     return re_opera_old.test(ua) ? re_opera_old : re_opera_new;
   }],
   ["oupeng", /\boupeng\/([0-9.]+)/],
@@ -371,7 +311,7 @@ var BROWSER = [
     return /\bversion\/([0-9.]+(?: beta)?)/;
   }],
   ["blackberry", function(ua){
-    var m = ua.match(re_blackberry_10) ||
+    const m = ua.match(re_blackberry_10) ||
       ua.match(re_blackberry_6_7) ||
       ua.match(re_blackberry_4_5);
     return m ? {version: m[1]} : "blackberry";
@@ -383,154 +323,10 @@ var BROWSER = [
   ["nokia", /\bnokiabrowser\/([0-9.]+)/],
 ];
 
-// UserAgent Detector.
-// @param {String} ua, userAgent.
-// @param {Object} expression
-// @return {Object}
-//    返回 null 表示当前表达式未匹配成功。
-function detect(name, expression, ua){
-  var expr = isFunction(expression) ? expression.call(null, ua) : expression;
-  if(!expr){ return null; }
-  var info = {
-    name: name,
-    version: NA_VERSION,
-    codename: "",
-  };
-  var t = toString(expr);
-  if(expr === true){
-    return info;
-  }else if(t === "[object String]"){
-    if(ua.indexOf(expr) !== -1){
-      return info;
-    }
-  }else if(isObject(expr)){ // Object
-    if(expr.hasOwnProperty("version")){
-      info.version = expr.version;
-    }
-    return info;
-  }else if(expr.exec){ // RegExp
-    var m = expr.exec(ua);
-    if(m){
-      if(m.length >= 2 && m[1]){
-        info.version = m[1].replace(/_/g, ".");
-      }else{
-        info.version = NA_VERSION;
-      }
-      return info;
-    }
-  }
-}
-
-var na = {name:"na", version:NA_VERSION};
-// 初始化识别。
-function init(ua, patterns, factory, detector){
-  var detected = na;
-  each(patterns, function(pattern){
-    var d = detect(pattern[0], pattern[1], ua);
-    if(d){
-      detected = d;
-      return false;
-    }
-  });
-  factory.call(detector, detected.name, detected.version);
-}
-
-// 解析 UserAgent 字符串
-// @param {String} ua, userAgent string.
-// @return {Object}
-var parse = function(ua){
-  ua = (ua || "").toLowerCase();
-  var d = {};
-
-  init(ua, DEVICES, function(name, version){
-    var v = parseFloat(version);
-    d.device = {
-      name: name,
-      version: v,
-      fullVersion: version,
-    };
-    d.device[name] = v;
-  }, d);
-
-  init(ua, OS, function(name, version){
-    var v = parseFloat(version);
-    d.os = {
-      name: name,
-      version: v,
-      fullVersion: version,
-    };
-    d.os[name] = v;
-  }, d);
-
-  var ieCore = IEMode(ua);
-
-  init(ua, ENGINE, function(name, version){
-    var mode = version;
-    // IE 内核的浏览器，修复版本号及兼容模式。
-    if(ieCore){
-      version = ieCore.engineVersion || ieCore.engineMode;
-      mode = ieCore.engineMode;
-    }
-    var v = parseFloat(version);
-    d.engine = {
-      name: name,
-      version: v,
-      fullVersion: version,
-      mode: parseFloat(mode),
-      fullMode: mode,
-      compatible: ieCore ? ieCore.compatible : false,
-    };
-    d.engine[name] = v;
-  }, d);
-
-  init(ua, BROWSER, function(name, version){
-    var mode = version;
-    // IE 内核的浏览器，修复浏览器版本及兼容模式。
-    if(ieCore){
-      // 仅修改 IE 浏览器的版本，其他 IE 内核的版本不修改。
-      if(name === "ie"){
-        version = ieCore.browserVersion;
-      }
-      mode = ieCore.browserMode;
-    }
-    var v = parseFloat(version);
-    d.browser = {
-      name: name,
-      version: v,
-      fullVersion: version,
-      mode: parseFloat(mode),
-      fullMode: mode,
-      compatible: ieCore ? ieCore.compatible : false,
-    };
-    d.browser[name] = v;
-  }, d);
-  return d;
+module.exports = {
+  device: DEVICES,
+  os: OS,
+  browser: BROWSER,
+  engine: ENGINE,
+  re_msie: re_msie,
 };
-
-
-// NodeJS.
-if(typeof process === "object" && process.toString() === "[object process]"){
-
-  // 加载更多的规则。
-  var morerule = module["require"]("./morerule");
-  [].unshift.apply(DEVICES, morerule.DEVICES || []);
-  [].unshift.apply(OS, morerule.OS || []);
-  [].unshift.apply(BROWSER, morerule.BROWSER || []);
-  [].unshift.apply(ENGINE, morerule.ENGINE || []);
-
-}else{
-
-  var userAgent = navigator.userAgent || "";
-  //var platform = navigator.platform || "";
-  var appVersion = navigator.appVersion || "";
-  var vendor = navigator.vendor || "";
-  external = win.external;
-
-  detector = parse(userAgent + " " + appVersion + " " + vendor);
-
-}
-
-// exports `parse()` API anyway.
-detector.parse = parse;
-
-module.exports = detector;
